Remove SymbolKey.GetComparer()
