+++ src/Workspaces/Core/Portable/xlf/WorkspacesResources.zh-Hant.xlf
@@ -12,6 +12,11 @@
         <target state="new">Refactoring Only</target>
         <note />
       </trans-unit>
+      <trans-unit id="Could_not_resolve_SymbolKey_because_it_does_not_contain_encoded_symbol_data">
+        <source>Could not resolve symbol key because it does not contain encoded symbol data.</source>
+        <target state="new">Could not resolve symbol key because it does not contain encoded symbol data.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="Symbol_0_is_not_from_source">
         <source>Symbol "{0}" is not from source.</source>
         <target state="translated">符號 "{0}" 非來自來源。</target>
